+++ src/ids-button/ids-button.js
@@ -9,6 +9,7 @@ import {
 
 import { IdsEventsMixin } from '../ids-base/ids-events-mixin';
 import { IdsThemeMixin } from '../ids-base/ids-theme-mixin';
+import { IdsTooltipMixin } from '../ids-base/ids-tooltip-mixin';
 import { IdsRenderLoopMixin, IdsRenderLoopItem } from '../ids-render-loop/ids-render-loop-mixin';
 
 import styles from './ids-button.scss';
@@ -55,13 +56,19 @@ const ICON_ALIGN = [
  * @mixes IdsRenderLoopMixin
  * @mixes IdsThemeMixin
  * @mixes IdsEventsMixin
+ * @mixes IdsTooltipMixin
  * @part button - the button element
  * @part icon - the icon element
  * @part text - the text element
  */
 @customElement('ids-button')
 @scss(styles)
-class IdsButton extends mix(IdsElement).with(IdsRenderLoopMixin, IdsEventsMixin, IdsThemeMixin) {
+class IdsButton extends mix(IdsElement).with(
+    IdsRenderLoopMixin,
+    IdsEventsMixin,
+    IdsThemeMixin,
+    IdsTooltipMixin
+  ) {
   constructor() {
     super();
     this.state = {};
@@ -113,6 +120,7 @@ class IdsButton extends mix(IdsElement).with(IdsRenderLoopMixin, IdsEventsMixin,
   }
 
   /**
+   * Figure out the classes
    * @private
    * @readonly
    * @returns {Array} containing classes used to identify this button prototype