+++ src/ids-container/ids-container.js
@@ -57,11 +57,13 @@ class IdsContainer extends mix(IdsElement).with(IdsEventsMixin, IdsThemeMixin) {
    */
   set scrollable(value) {
     if (stringUtils.stringToBool(value)) {
-      this.setAttribute('scrollable', value.toString());
+      this.setAttribute('scrollable', 'true');
+      this.container.setAttribute('scrollable', 'true');
       return;
     }
 
     this.setAttribute('scrollable', 'false');
+    this.container.setAttribute('scrollable', 'false');
   }
 
   get scrollable() { return this.getAttribute('scrollable') || 'true'; }